Rename “OS X” to the newer “macOS” convention
Summary:
```
Since 2016, with macOS 10.12 Sierra, Mac OS X has been renamed in macOS.
It would be a nice if Bitcoin's macOS build instructions follow this
naming convention to avoid misunderstandings.
```

Backport of core PR13366
https://github.com/bitcoin/bitcoin/pull/13366/files

Depends on D3910.
Includes name changes to the gitian documentation which are not part of
the original PR.

Test Plan: Read it.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3912