Bump automatic replay protection to May 2019 upgrade
Summary:
This returns main chain to ForkID of 0 after the Nov 2018
upgrade, and implements automatic replay protection for
the May 2019 upgrade.

For reference see last HF:
https://reviews.bitcoinabc.org/D1199

Test Plan:
Replay protection is covered by unit tests, however
it is very important to manually double check that the timestamp
for activation is set correctly.

Reviewers: deadalnix, schancel, jasonbcox, #bitcoin_abc

Reviewed By: deadalnix, schancel, jasonbcox, #bitcoin_abc

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1612