RPC: Add new getzmqnotifications method.
Summary:
```
This adds a new RPC method "getzmqnotifications", which returns
information about all active ZMQ notification endpoints.  This is useful
for software that layers on top of bitcoind, so it can verify that
ZeroMQ is enabled and also figure out where it should listen.
```

Backport of core PR13570 and PR13645:
https://github.com/bitcoin/bitcoin/pull/13570/files
https://github.com/bitcoin/bitcoin/pull/13645/files

Test Plan:
  make check

Run 3 times:
  test_runner.py rpc_zmq
 - First time with zmq all on, should pass
 - Second time after uninstalling python3-zmq, should skip (reinstall
after :))
 - Third time after disabling zmq (with configure or cmake), should skip

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4447