Add aarch64 qt depends support for cross compiling bitcoin-qt
Summary:
Backport of core PR13696:
https://github.com/bitcoin/bitcoin/pull/13696/files

This is a follow-up for this comment:
https://reviews.bitcoinabc.org/D4356?id=14046#inline-27298

Depends on D4356.

Test Plan:
Rebuild the aarch64 depends.

Cross build with CMake.
Run the binary on an aarch64 platform and check it works (tested on a
Rpi3 running Mint 64).

Run the linux gitian build.
Run the binary on an aarch64 platform and check it works (tested on a
Rpi3 running Mint 64).

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4461