[qa] fix flaky abc-magnetic-anomaly-activation.py
Summary:
Currently, this test has a tranasction generator for a specific size.
It slowly grows the transaction padding to ensure it is exactly 101 bytes.
However, if the penultimate iteration of the loop generates a padding with
a leftmost bit of zero, and the ultimate run generates a padding with a
leftmost bit of 1, the test will fail.

This commit ensure the leftmost bit is always 1.

Test Plan:
Run in a loop for 100 runs:

  ./test/functional/test_runner.py abc-magnetic-anomaly-activation.py

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, deadalnix, jasonbcox

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1700