Increase INVENTORY_BROADCAST_MAX to 1170 (based on 32MB max message size). This allows ABC to advertise transactions at up to 234 tx/sec. For a node with 50 peers and ~50 bytes per tx (including INV overhead), this would consume about 585 kB/s of...
Summary: ...upstream bandwith for INVs, plus an unspecified and unpredictable amount of bandwidth for the transactions themselves.

Test Plan: test_bitcoin works. Running with debug=net appears to work as expected. I have not tested this with higher tx throughput than is currently available on mainnet, as no such testnet is currently available.

Reviewers: deadalnix, #bitcoin_abc

Reviewed By: deadalnix, #bitcoin_abc

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1784