Add callsign to .arcconfig for proper repo identification
Summary:
If you have remotes configured for a branch, `arc` gets confused when trying
to find the configuration for staging areas. This change explicitly sets the
repository callsign in `.arcconfig` which arc uses to look up the appropriate
staging area.

Test Plan: Push a commit and see if it goes to staging properly

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1814