Fixed bug in UnwindBlock where child blocks do not have proper parent flags set
Summary:
Fixed a typo introduced in my fix from 24d17379f

Important note: There is unintuitive behavior now documented in the test in the form of NOTE comments.  The order of calling `parkblock` and `invalidateblock` impacts the chaintip status.

Test Plan:
test_runner.py abc-parkedchain
The new test fails without the fix and passes with it.

Reviewers: deadalnix, schancel, #bitcoin_abc, Fabien

Reviewed By: deadalnix, #bitcoin_abc, Fabien

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1987