[tests] fix cross-platform sprintf issue in blockindex_tests
Summary:
As per title.  Tinyformat does not print the same on every platform w/r/t pointers.
It does not seem follow the libc specification on Darwin.  This patch fixes the issue
by using %p in both the test, and the ToString() method.

Additionally, use BOOST_CHECK_EQUAL so proper failure strings are printed in
the future.

Test Plan:
  make VERBOSE=1 check

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1997