Add a way to stop scheduleEvery instances.
Summary:
Once a task is scheduled to happen on a regular basis, there are no ways to stop it. This diff fixes it.

It does so by making the scheduled function return a bool, which will determine if it is to be executed again.

Test Plan: Add a test case.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: Fabien

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D2011