Cleanup redundant function in abc-p2p-compactblocks
Summary:
See title. Rehashing the txs is unnecessary because make_conform_to_ctor does this for us.

Co-authored-by: Shammah Chancellor <shammah.chancellor@bitcoinabc.org>

Test Plan: Test passes up to the point where it already fails, so there's that I guess? (we're still working on the test failure for this file; cleaning up helps us focus and move forward)

Reviewers: deadalnix, schancel, #bitcoin_abc

Reviewed By: schancel, #bitcoin_abc

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D2101