wallet: Avoid leaking nLockTime fingerprint when anti-fee-sniping
Summary:
```
The wallet sets the locktime to the current height of our active chain.
This is fine, as long as our node is connected to other nodes. However,
when we fall back and get stuck at a particular height (e.g. taking the
wallet offline), the same (potentially unique) locktime is used for all
transactions. This makes it easier for passive observers to cluster
transactions by wallet.
```

Backport of core [[https://github.com/bitcoin/bitcoin/pull/15039 | PR15039]].

Test Plan:
  ninja check check-functional-extended

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5560