Integrate the string format linter with arcanist
Summary:
The linter can be run with `arc lint` to check all the files in a diff.
The bash script now takes a single file as an argument to make it
compliant with arcanist linter expectations.

Depends on D2025 and D2225

Test Plan:
Modify some files with a string format error and run `arc lint`.
Check that the error is catched by the linter.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2218