Fix Clang -Wcast-align warnings
Summary:
Remove some c-style cast that raised warnings on clang due to change in
the alignment requirement. Using `reinterpret_cast<>()` will
prevent these warnings.

Test Plan:
Using clang as a compiler:
  make check
Ensure that there is no -Wcast-align warning from the compiler on the files that are updated in this diff

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2256