Improve check-doc.py argument counting output
Summary:
The used arguments count did not reflect exactly the actual number of
arguments used because it did not take the SET_DOC_OPTIONAL list into
account.
Also add the number of explicitely undocumented argument to the output.

Depends on D2299

Test Plan:
  ./contrib/devtools/check-doc.py
Compare the outputs before and after the patch.
Check the count is augmented by the correct number of arguments

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2303