Fix unknown arguments from check-doc.py
Summary:
The `-zmq*` arguments are not catched by the regex because they are
using a totally different pattern to get the arguments. A false positive
set has been added to check-doc.py to handle these special cases.

The `-nodebug` entry is an alias for `-debug=0`. Starting with "-no" it
is interpreted as a boolean with "false" value for the key "debug".
Boolean arguments are actully strings "0" for "false" and "1" for
"true", hence `-nodebug` is equivalent to `-debug=0`

The `-fuzzmessagestest` is no longer used and can be removed from the
documentation.

Depends on D2303

Test Plan:
  make
  ./contrib/devtools/check-doc.py
Check there is no unknown argument in the output (empty set)
Check that the count of args used = (count of args documented + count of
args that don't need documentation)

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2304