Bump automatic replay protection to Nov 2019 upgrade
Summary:
This returns main chain to ForkID of 0 after the May 2019 upgrade and implements automatic replay protection for the Nov 2019 upgrade.

For reference see last bump in D1612

(original author of diff: dagurval)

Test Plan: Replay protection is covered by unit tests, however it is very important to manually double check that the timestamp for activation is set correctly.

Reviewers: #bitcoin_abc, deadalnix, Mengerian, dagurval

Reviewed By: #bitcoin_abc, deadalnix, Mengerian

Subscribers: markblundeberg, Mengerian, teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2376