Bump minimum Qt version to 5.5.1
Summary:
This version is already required as per our dependencies.md
documentation but is not enforced by the build system.

Backport of core PR15393:
https://github.com/bitcoin/bitcoin/pull/15393/files

Depends on D4502.

Note to reviewers:
Because >= 5.4 is already enforced de facto, and our documentation
already requires >=5.5.1, I didn't add the release notes to the diff.

Test Plan:
  make check # autotools, need to re-run autogen.sh
  ninja check

Change the minimum version to 5.15.1 and check the build systems bail
out with an meaningful error. This works with cmake but probably won't
work with autotools, intended that the version check output is never
tested.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D4507