Integrates the python format linter with arcanist
Summary:
Running `arc lint` or `arc diff` will run the linter-python-format.py
script on all the python files in the diff (excepted the linter itself
to avoid erasing the doctest).
It will automatically propose to amend the diff with the new format
applied to file, in the same way that autopep8 or clang-format do.

Depends on D2557

Test Plan:
Add a comment in `test/functional/test_framework/messages.py` then run
`arc lint`.
Check that arcanist offer to amend 31 format replacements for the file.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2558