Add a test for signrawtransaction with various sighashes parameters
Summary:
While reviewing D2007 @jasonbcox raised that there was not test for the
various errors thrown when `signrawtransaction` is called with an
invalid sighash parameter.
This diff adds a functional test case which check the output of
`signrawtransaction` against valid sighashes, invalid sighashes and
sighashes missing the FORKID flag.

Test Plan:
  ./test/functional/test_runner.py rpc_signrawtransaction

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: teamcity, schancel, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D2560