Remove unused enable/disable_mocktime() from test_framework.py
Summary:
This is a follow-up for D2628 and D2629.
It appeared that these changes are part of a core PR (PR14738).
This diff backports the remaining parts from the original PR.
Note that the test is renamed in the original PR, this is a change from
core PR12953 that could be backported later.

Partial backport of core PR14738
Depends on D2635

Test Plan:
  ./test/functional/test_runner.py --extended

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2636