[LINTER] Integrate lint-locale-dependence.sh into arcanist
Summary:
Change the linter to accept a file as an argument, and always return 0
to avoid arcanist to throw a script execution error.
Add a php class to interpret the linter output and display linting error
messages through `arc lint`.

Depends on D2704

Test Plan:
  arc lint

Comment any of the known exception in the `linter-locale-dependence.sh`
file and add a comment in the corresponding source file. Check that `arc
lint` displays errors for the source file.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox, teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2705