[LINTER] Display the line number in the lint-locale-dependence.sh linter
Summary:
This makes the output to look like this:
  <source file>:<line nunmber>:<code snippet>
Having the line number makes it easier to locate the code line and make
sure all the issues are fixed. This is especially true in this case
where it is likely to have multiple times the same line in the file.

Depends on D2698

Test Plan:
  ./test/lint/lint-locale-dependence.sh
Check the script outputs not error (no output)

Comment any of the known exceptions in the `lint-locale-dependence.sh`
script.
  ./test/lint/lint-locale-dependence.sh
Check the output contains the line number

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2704