[backport#15646] [tests] Add test for wallet rebroadcasts
Summary:
The existing wallet_resendwallettransactions.py test only tests the
resendwallettransactions RPC. It does not test whether transactions are
actually rebroadcast, or whether the rebroadcast logic is called on a
timer.

This commit updates the test to not use the resendwallettransactions RPC and
test that transactions are rebroadcast on a timer.

---

This is a backport of Core [[https://github.com/bitcoin/bitcoin/pull/15646 | PR15646]]

Test Plan:
  cmake .. -GNinja
  ./test/functional/test_runner.py wallet_resendwallettransactions.py

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D5947