[backport#15638] [build] Move CheckTransaction from lib_server to lib_consensus
Summary:
CheckTransaction (N.f.B: CheckCoinbase, CheckTransactionCommon and CheckRegularTransaction
in our case ) is a context-free function that does not require access to the blockchain
or mempool. Move it from src/consensus/tx_verify in lib_server to a new unit
src/consensus/tx_check in lib_consensus so that it can be called by non-server libraries.

[[https://github.com/bitcoin/bitcoin/pull/15638/commits/fdf8888b6f0c63e8a4cb1459752625e642d6a4dd | fdf8888b6f0c63e8a4cb1459752625e642d6a4dd]] (John Newbery)

---

This is a partial backport of Core [[https://github.com/bitcoin/bitcoin/pull/15638 | PR15638]]

Test Plan:
  ninja check check-functional

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5961